api/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.10
django-cors-middleware==1.5.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-rest-swagger==2.2.0
djangorestframework==3.10.3
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
openapi-codec==1.3.2
pytz==2019.3
requests==2.23.0
responses==0.10.11
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
